Split files back out
