+++ src/index.ts
@@ -1,15 +1,14 @@
-import { program } from 'commander';
 import { config } from 'process';
 
 const chokidar = require('chokidar');
 const color = require('picocolors');
 const { Command } = require('commander');
 const fs = require('fs');
-const fg = require('fast-glob');
 const path = require('path');
-const picomatch = require('picomatch');
 const postcss = require('postcss');
 const sass = require('sass');
+
+const parser = require('./parser');
 const ver = require('../package.json').version;
 
 const sheetloaf = new Command();
@@ -69,7 +68,7 @@ function init(source) {
 	if (usingStdin === true) {
 		render(source);
 	} else {
-		expandGlob(source[0].split(','), function (entries) {
+		parser.expandGlob(source[0].split(','), function (entries) {
 			entries.forEach(function (filename) {
 				if (path.basename(filename).charAt(0) !== '_') {
 					render(filename);
@@ -95,7 +94,7 @@ function watch(source) {
 			.on('change', (changed) => {
 				console.log(`File changed: ${changed}`);
 
-				expandGlob(source[0].split(','), function (entries) {
+				parser.expandGlob(source[0].split(','), function (entries) {
 					entries.forEach(function (filename) {
 						if (path.basename(filename).charAt(0) !== '_') {
 							render(filename);
@@ -106,7 +105,7 @@ function watch(source) {
 			.on('add', (added) => {
 				console.log(`File added: ${added}`);
 
-				expandGlob(source[0].split(','), function (entries) {
+				parser.expandGlob(source[0].split(','), function (entries) {
 					entries.forEach(function (filename) {
 						if (path.basename(filename).charAt(0) !== '_') {
 							render(filename);
@@ -268,61 +267,6 @@ function generateSassOptions(source: string, destination: string) {
 			sourceMapEmbed: sheetloaf.opts().sourceMap === false ? false : true,
 			includePaths: sheetloaf.opts().loadPath.split(',')
 		};
-	}
-}
-
-/**
- * Given an array of input sources, expand any globs and validate that inputs exist.
- *
- */
-function expandGlob(input: string[], callback, index: number = 0, expanded: string[] = []) {
-	if (index < input.length) {
-		let isGlob = false;
-		let isDir = false;
-		let isFile = false;
-
-		isGlob = picomatch.scan(input[index]).isGlob;
-
-		if (isGlob === false) {
-			try {
-				isDir = fs.lstatSync(path.normalize(input[index])).isDirectory();
-				isFile = fs.lstatSync(path.normalize(input[index])).isFile();
-			} catch (err) {
-				throw err;
-			}
-		}
-
-		if (isGlob || isFile) {
-			let files = fg
-				.sync(input[index], {
-					dot: true
-				})
-				.map((entry) => path.normalize(entry));
-			expanded.push(...files);
-
-			index = index + 1;
-			expandGlob(input, callback, index, expanded);
-		} else if (isDir) {
-			let dir = input[index];
-
-			fs.readdir(dir, (err, files) => {
-				if (err) {
-					throw err;
-				} else {
-					files.forEach((file) => {
-						let fullname = path.join(dir, file);
-						if (!fs.lstatSync(fullname).isDirectory()) {
-							expanded.push(fullname);
-						}
-					});
-
-					index = index + 1;
-					expandGlob(input, callback, index, expanded);
-				}
-			});
-		}
-	} else {
-		callback(expanded);
 	}
 }
