+++ test/samples/index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <title>Test Page</title>
+
+    <!-- Housekeeping -->
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <!-- Styles -->
+    <link rel="stylesheet" href="./render/file.css">
+</head>
+
+<body>
+    <div class="container">
+        <div><p>hello world</p></div>
+    </div>
+    <!-- END .container -->
+    <div class="testElement"></div>
+</body>
+
+</html>
