+++ test/samples/lib/_mixins.scss
@@ -1,0 +1,4 @@
+@mixin happyBg {
+    // a nice sunny yellow.
+    background-color: yellow;
+}
