+++ test/samples/postcss.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    plugins: [
+        require('autoprefixer'),
+        require('postcss-custom-properties')
+    ]
+}
