+++ test/samples/render/file-with-error.css
@@ -1,0 +1,1 @@
+body:before { content: 'Line 9: Error: expected "{".    } jasldfjkasj; kfjq;wkjf[23rjr23 dj lwdj ;LFKJAS;LDKJ AS;DLJF AS;LKDJFQ3 EDJ ]                 ';display: table;background-color:#cc0000;color:white;border-radius:5px;margin-bottom:5px;padding:5px;font-family:sans-serif}body:after { content: '/home/ben/Git Projects/sheetloaf/tests/samples/styles/file-with-error.scss';display: table;background-color:#0e70b0;color:white;border-radius:5px;padding:5px;margin-bottom: 5px;font-family:sans-serif}body * { display: none; }
