+++ test/samples/styles/_mixins.scss
@@ -1,0 +1,3 @@
+@mixin happyBg {
+    background-color: yellow;
+}
