+++ test/samples/styles/file-with-error.scss
@@ -1,0 +1,9 @@
+body {
+    background: green;
+    display: flex; 
+    display: grid;
+    backdrop-filter: blur(10px);
+    div {
+        color: red;
+    }
+} jasldfjkasj; kfjq;wkjf[23rjr23 dj lwdj ;LFKJAS;LDKJ AS;DLJF AS;LKDJFQ3 EDJ ]
