+++ test/samples/styles/file.scss
@@ -1,0 +1,16 @@
+@use 'mixins';
+@use 'partial';
+
+:root {
+    --color: green;
+}
+
+.container {
+    backdrop-filter: blur(10px);
+    background: var(--color);
+
+    div {
+        @include mixins.happyBg;
+        width: 20px;
+    }
+}
