--- tests/samples/lib/_mixins.scss
@@ -1,4 +1,0 @@
-@mixin happyBg {
-    // a nice sunny yellow.
-    background-color: yellow;
-}
