--- tests/samples/logs/error.log
@@ -1,6 +1,0 @@
-Error: expected "{".
-  ╷
-9 │ } jasldfjkasj; kfjq;wkjf[23rjr23 dj lwdj ;LFKJAS;LDKJ AS;DLJF AS;LKDJFQ3 EDJ ]
-  │              ^
-  ╵
-  stdin 9:14  root stylesheet
