--- tests/samples/render/file.css
@@ -1,29 +1,0 @@
-a {
-  color: red;
-  border-radius: 5px;
-}
-
-.testElement {
-  margin-top: 25px;
-  height: 300px;
-  width: 300px;
-  background: blue;
-  position: absolute;
-  top: 50px;
-  left: 50px;
-}
-
-:root {
-  --color: green;
-}
-
-.container {
-  -webkit-backdrop-filter: blur(10px);
-          backdrop-filter: blur(10px);
-  background: green;
-  background: var(--color);
-}
-.container div {
-  background-color: yellow;
-  width: 20px;
-}
