--- tests/samples/styles/_mixins.scss
@@ -1,3 +1,0 @@
-@mixin happyBg {
-    background-color: yellow;
-}
