--- tests/samples/styles/_partial.scss
@@ -1,15 +1,0 @@
-a {
-    color: red;
-    -webkit-border-radius: 5px;
-    border-radius: 5px;
-}
-
-.testElement {
-    margin-top: 25px;
-    height: 300px;
-    width: 300px;
-    background: blue;
-    position: absolute;
-    top: 50px;
-    left: 50px;
-}
